backend/requirements.txt
asgiref==3.3.4
autopep8==1.5.7
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
Django==3.2.3
django-appconf==1.0.4
django-environ==0.4.5
django-googledrive-storage==1.5.2
django-ipware==3.0.2
djangorestframework==3.12.4
drf-recaptcha==2.0.4
google-api-core==1.28.0
google-api-python-client==2.6.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
oauth2client==4.1.3
packaging==20.9
Pillow==8.2.0
protobuf==3.17.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
